--- a/src/plugins/ui/filter_evaluation.ts
+++ b/src/plugins/ui/filter_evaluation.ts
@@ -197,9 +197,10 @@ export class FilterEvaluationPlugin extends UIPlugin {
       if (hiddenRows.has(filter.zoneWithHeaders.top)) continue;
       const filteredValues = this.filterValues[sheetId]?.[filter.id]?.map(toLowerCase);
       if (!filteredValues || !filter.filteredZone) continue;
+      const filteredValuesSet = new Set(filteredValues);
       for (let row = filter.filteredZone.top; row <= filter.filteredZone.bottom; row++) {
         const value = this.getCellValueAsString(sheetId, filter.col, row);
-        if (filteredValues.includes(value)) {
+        if (filteredValuesSet.has(value)) {
           hiddenRows.add(row);
         }
       }