@robodoo 2025-03-20
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/o-spreadsheet/pull/5977.png)](https://mergebot.odoo.com/odoo/o-spreadsheet/pull/5977)